(Promoted)Checker And Encapsulation?
I have added the kingChecker which should be the promoted piece and the logic backwards should be there. If you wondering why I was getting rid of the board in the Checker parameter is because it would be better encapsulation if only the board manages that and the checkers give me of a general rule set. I don't think it would affect the inversion but I guess you can yell at me if it does. Also I haven't completed all the classes so it might no compile. I just put the logic